pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "job-application-tracker-api"
version = "0.1.0"
description = "Production-style FastAPI backend for tracking job applications."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13.3",
    "asyncpg>=0.29.0",
    "bcrypt>=4.0.1,<4.1.0",
    "celery[redis]>=5.4.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.1",
    "pydantic-settings>=2.5.2",
    "python-jose[cryptography]>=3.3.0",
    "redis>=5.0.8",
    "sqlalchemy[asyncio]>=2.0.35",
    "uvicorn[standard]>=0.30.6",
]

[project.optional-dependencies]
dev = [
    "aiosqlite>=0.20.0",
    "black>=24.8.0",
    "httpx>=0.27.2",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.9",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC", "N"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore:'crypt' is deprecated:DeprecationWarning",
]
